[Windows Store 8.x] Locks the swapChain to the native orientation
Tells to the system to not rotate the swapChain automatically.
The program is responsible to rotate camera,UI, Input, etc.
The behavior is somewhat similar to when using
DrawingSurfaceBackgroundGrid  on WP8.
Although, touch input behave differently between the two platforms,
on WP8 you have to rotate touch using the same 2D rotation matrix you
would use for drawing UI elements. On win8 touch input is allready
rotated by the system.